Updated readme.md to include new additionOptions argument for new features.
New features include ability to specify whether or not labels, conservation scores, and references sequences should be visible. Also, options were added for specifying tick size and whether or not grid lines should be present at the tick lines. Zoom in feature was added with specification of a zoom range. Lastly, the option to add a second track of mutations was added to the plotProteinWithConservation.R script.